% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPerturbationNode.R
\name{AddPerturbationNode}
\alias{AddPerturbationNode}
\title{\code{AddPerturbationNode}}
\usage{
AddPerturbationNode(network)
}
\arguments{
\item{network}{The original prior knowledge network}
}
\value{
A list of updated network with perturbation nodes and re-assigned list of input to include only the interactions from the perturbation node.
}
\description{
Introduces a perturbation node connecting periphery nodes without a target in the prior knowledge network.
}
