% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_perturbation_node.R
\name{addPerturbationNodes}
\alias{addPerturbationNodes}
\title{Introduces a perturbation node connecting periphery nodes without a target in 
the prior knowledge network.}
\usage{
addPerturbationNodes(priorKnowledgeNetwork)
}
\arguments{
\item{priorKnowledgeNetwork}{data.frame with priorKnowledgeNetwork with source, interaction, target columns}
}
\value{
data.frame with prior knowledge network with added perturbations
}
\description{
Introduces a perturbation node connecting periphery nodes without a target in 
the prior knowledge network.
}
\author{
Panuwat Trairatphisan, 2020
}
\keyword{internal}
