% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPROGENyScores.R
\name{assignPROGENyScores}
\alias{assignPROGENyScores}
\title{\code{assignPROGENyScores}}
\usage{
assignPROGENyScores(progeny = progeny, progenyMembers = progenyMembers,
  id = "uniprot")
}
\arguments{
\item{progeny}{Contains the progeny scores as obtained from \code{runPROGENy}.}

\item{progenyMembers}{Contains the list of members for each PROGENy pathway.}

\item{id}{Contains the members identifiers (default: uniprot).}
}
\value{
This function is used to account for the PROGENy scores in the objective function.
}
\description{
\code{assignPROGENyScores}
}
