% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataMatrix.R
\name{buildDataMatrix}
\alias{buildDataMatrix}
\title{\code{buildDataMatrix}}
\usage{
buildDataMatrix(data = data, pknList = pknList, inputs = inputs)
}
\arguments{
\item{data}{Contains the measured data.}

\item{pknList}{Contains the background network which serves as a prior knowledge and which we train.}

\item{inputs}{Contains the list of targets as inputs.}
}
\value{
This function returns the data matrix containing the data for running CARNIVAL and a set of identifiers for Targets, Measured and Un-measured nodes.
}
\description{
\code{buildDataMatrix}
}
