% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_visNet.R
\name{carnival_visNet}
\alias{carnival_visNet}
\title{carnival_visNet}
\usage{
carnival_visNet(evis, nvis, mapIDs = NULL)
}
\arguments{
\item{evis}{dataframe with summary edge information from CARNIVAL}

\item{nvis}{dataframe with summary node information from CARNIVAL}

\item{mapIDs}{dataframe with values of hgnc_symbol and uniprot_id to add hgnc as labels of the nodes}
}
\value{
visNetwork object with the visualization of the network
}
\description{
Visualization of CARNIVAL results using visNetwork
}
\examples{
# Sample network data
network_data <- data.frame(
  Node1 = c("A", "A", "B", "B", "C", "D", "E"),
  Sign = c(1, -1, 1, -1, 1, 1, -1),
  Node2 = c("B", "C", "C", "D", "E", "F", "F"),
  Weight = rep(100,7)
)

node_attribute <- data.frame(
  Node = c("A","B","C","D","E","F"),
  NodeType = c("P","","","","M","M"),
  ZeroAct = rep(0,6),
  UpAct = c(100,100,0,0,100,0),
  DownAct = c(0,0,100,100,0,100),
  AvgAct = c(100,100,-100,-100,100,-100)
)
carnival_visNet(network_data, node_attribute)
}
