% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_solver_param.R
\name{checkCarnivalOptions}
\alias{checkCarnivalOptions}
\title{Checks options provided for CARNIVAL}
\usage{
checkCarnivalOptions(carnivalOptions)
}
\arguments{
\item{carnivalOptions}{all available carnival options}
}
\description{
Checks options provided for CARNIVAL
}
\keyword{internal}
