% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{checkData}
\alias{checkData}
\title{Checks the input data for correctness.}
\usage{
checkData(
  perturbations = NULL,
  measurements,
  priorKnowledgeNetwork,
  weights = NULL
)
}
\arguments{
\item{perturbations}{}

\item{measurements}{}

\item{priorKnowledgeNetwork}{}

\item{weights}{}
}
\value{

}
\description{
Checks the input data for correctness.
}
\author{
Enio Gjerga, Olga Ivanova, Attila Gabor, 2020-2021
}
\keyword{internal}
