% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{checkOptionsValidity}
\alias{checkOptionsValidity}
\title{Checks if provided option names are valid.}
\usage{
checkOptionsValidity(solver = getSupportedSolvers()$lpSolve, ...)
}
\arguments{
\item{solver}{one of the solvers available from getSupportedSolvers().}

\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Checks if provided option names are valid.
}
\examples{
checkOptionsValidity(solver="lpSolve")
}
