% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables_all.R
\name{create_variables_all}
\alias{create_variables_all}
\title{\code{create_variables_all}}
\usage{
create_variables_all(pknList = pknList, dataMatrix = dataMatrix)
}
\arguments{
\item{pknList}{Contains the background network which serves as a prior knowledge and which we train.}

\item{dataMatrix}{Contains the matrix which stores the data for running CARNIVAL and a set of identifiers for Targets, Measured and Un-measured nodes.}
}
\value{
This function returns the identifiers of all the variables used in the ILP formulation together with an explanation about the meaning of each of them. Also it returns a list of useful identifiers.
}
\description{
\code{create_variables_all}
}
