% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{createInternalDataRepresentation}
\alias{createInternalDataRepresentation}
\title{Creates internal data representation - variables for ILP solvers, on the 
basis of provided preprocessed data.}
\usage{
createInternalDataRepresentation(dataPreprocessed, newDataRepresentation = T)
}
\arguments{
\item{dataPreprocessed}{list containing preprocessed priorKnowledgeNetwork, 
measurements, weights (if provided), perturbations (if provided).}

\item{newDataRepresentation}{TRUE by default. For debugging with the old 
data representation, put to FALSE.}
}
\value{
variables for the new data representation or
data vector (containing preprocessed information on measurement) 
and variables for the old data representation (CARNIVAL v.<2)
}
\description{
Creates internal data representation - variables for ILP solvers, on the 
basis of provided preprocessed data.
}
\keyword{internal}
