% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{defaultCbcSolveCarnivalOptions}
\alias{defaultCbcSolveCarnivalOptions}
\title{Sets default CARNIVAL options for cbc.}
\usage{
defaultCbcSolveCarnivalOptions(...)
}
\arguments{
\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Sets default CARNIVAL options for cbc.
}
\examples{
#defaultCbcSolveCarnivalOptions()
}
