% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_carnival_options.R
\name{defaultCplexCarnivalOptions}
\alias{defaultCplexCarnivalOptions}
\title{default_CARNIVAL_options}
\usage{
defaultCplexCarnivalOptions(solverPath = "")
}
\arguments{
\item{solverPath}{Path to executable cbc/cplex file - default set to NULL, in
which case the solver from lpSolve package is used.}

\item{solver}{Solver to use: lpSolve/cplex/cbc (Default set to lpSolve).}

\item{timelimit}{CPLEX/Cbc parameter: Time limit of CPLEX optimisation in
seconds (default set to 3600).}

\item{mipGAP}{CPLEX parameter: the absolute tolerance on the gap between the
best integer objective and the objective of the best node remaining. When this
difference falls below the value of this parameter, the linear integer
optimization is stopped (default set to 0.05)}

\item{poolrelGAP}{CPLEX/Cbc parameter: Allowed relative gap of accepted
solution comparing within the pool of accepted solution (default: 0.0001)}

\item{limitPop}{CPLEX parameter: Allowed number of solutions to be generated
(default: 500)}

\item{poolCap}{CPLEX parameter: Allowed number of solution to be kept in the
pool of solution (default: 100)}

\item{poolIntensity}{CPLEX parameter: Intensity of solution searching
(0,1,2,3,4 - default: 4)}

\item{betaWeight}{Objective function: weight for node penalty (defaul: 0.2)}

\item{threads}{CPLEX parameter: Number of threads to use
default: 0 for maximum number possible threads on system}

\item{dirName}{Specify directory name to store results. by default set to
NULL}
}
\value{
returns a list with all possible options implemented in CARNIVAL.
see the documentation on \code{\link{CARNIVAL::runCARNIVAL}}.
}
\description{
generates default CARNIVAL options.
}
