% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{defaultCplexSpecificOptions}
\alias{defaultCplexSpecificOptions}
\title{Sets default options from cplex documentation.}
\usage{
defaultCplexSpecificOptions(...)
}
\arguments{
\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Sets default options from cplex documentation.
}
