% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{defaultLpSolveCarnivalOptions}
\alias{defaultLpSolveCarnivalOptions}
\title{Sets default CARNIVAL options for lpSolve.}
\usage{
defaultLpSolveCarnivalOptions(...)
}
\arguments{
\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Sets default CARNIVAL options for lpSolve.
}
\examples{
defaultLpSolveCarnivalOptions()
}
