% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichCARNIVAL.R
\name{enrichCARNIVAL}
\alias{enrichCARNIVAL}
\title{\code{enrichCARNIVAL}}
\usage{
enrichCARNIVAL(Result_dir = "Results_CARNIVAL", universeFile = NULL,
  datasource = "kegg", topEnrichAll = 40, pathwayfilter = T,
  pValSig = 0.05)
}
\arguments{
\item{Result_dir}{The Result CARNIVAL folder to perform enrichment analyses}

\item{universeFile}{The file containing the list of nodes/genes (with the header(s) 'genesymbol' +/- 'uniprot') to be considered as universe in enrichment analyses}

\item{datasource}{The data source on MSigDB C2 branch: c("kegg":default,"biocarta","reactome","allc2")}

\item{topEnrichAll}{Number of top enrichmed pathways to be plotted (default: 40)}

\item{pathwayfilter}{Filtering pathways which are related to cancers, infections and diseases to highlight cellular processes}

\item{pValSig}{Significant p-value to be labelled on the plotted figures}
}
\value{
Enriched pathways with p-value written into files and figures showing top enriched pathways for combined Up/Down direction and separated
}
\description{
This function performs the enrichment of CARNIVAL nodes with over-representation analysis using the curated gene set from MSigDB (C2 branch)
The enrichment can be performed using the KEGG, Biocarta or Reactome subsets or using the whole curated C2 geneset.
Note: The universe of the geneset should be all the nodes in the prior knowledge network. Here the nodes from Omnipath are used by default.
}
