% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResult.R
\name{exportResult}
\alias{exportResult}
\title{\code{exportResult}}
\usage{
exportResult(cplexSolutionFileName = cplexSolutionFileName,
  variables = variables, conditionIDX = conditionIDX,
  pknList = pknList, dir_name = dir_name, inputs = inputs,
  measurements = measurements, Export_all = Export_all,
  writeIndividualResults = F)
}
\arguments{
\item{cplexSolutionFileName}{Path to the cplex solution file (XML)}

\item{variables}{The list of mapping indices of LP variables}

\item{conditionIDX}{The number of experimental condition to be processed}

\item{pknList}{The provided prior knowledge network}

\item{dir_name}{The name of directory to store results}

\item{inputs}{The list of known or potential target of perturbation}

\item{measurements}{The discretised observations (here transcription factor activities) of values [-1,0,1]}

\item{Export_all}{An option to define whether all detailed mapppd LP variables will be written as individual files}

\item{writeIndividualResults}{An option to define whether the results of individual solutions will be written; if FALSE, only the global combined solution will be written}
}
\value{
Output files of ILP solutions and a list of networks and node activities to be written into figures
}
\description{
Extract and export the optimisation results from the cplex solution file (XML) as files and variables for further plotting functions
}
