% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files2res.R
\name{files2res}
\alias{files2res}
\title{\code{files2res}}
\usage{
files2res(counterlist)
}
\arguments{
\item{counterlist}{Indices of model solutions to be read-in}
}
\value{
A list of variables for plotting functions
}
\description{
Read in a list of exported/written network and node activities files into a variable for plotting
}
