% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_measfile.R
\name{generate_measfile}
\alias{generate_measfile}
\title{\code{generate_measfile}}
\usage{
generate_measfile(measurements, topnumber = NULL,
  write2folder = "./measurements")
}
\arguments{
\item{measurements}{A dataframe describing the measurements (e.g. inferred TF activities). Columns/colnames should correspond to perturbations, row/rownames to measurements.}

\item{topnumber}{If given, only the top number of measurements will be written out.}

\item{write2folder}{Path to outputfolder}
}
\value{
A n*m dataframe describing the inferred TF activities, where n is the number of Tfs and m the number of conditions.
}
\description{
This function generates the measurement file for each condition which is required as one of the input file
}
