% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_run_carnival.R
\name{generateLpFileCarnival}
\alias{generateLpFileCarnival}
\title{\code{generateLpFileCarnival}}
\usage{
generateLpFileCarnival(
  perturbations,
  measurements,
  priorKnowledgeNetwork,
  pathwayWeights = NULL,
  solver = supportedSolvers$lpSolve,
  solverPath = "",
  carnivalOptions = defaultLpSolveCarnivalOptions()
)
}
\arguments{
\item{perturbations}{(optional, if inverse CARNIVAL flavour is used further) vector of targets of perturbations.}

\item{measurements}{vector of the measurements (i.e. DoRothEA/VIPER normalised
enrichment scores)}

\item{priorKnowledgeNetwork}{data frame of the prior knowledge network}

\item{pathwayWeights}{(optional) vector of the additional weights: e.g. PROGENy pathway
score or measured protein activities.}
}
\value{
data frame of all variables for ILP formulation.
}
\description{
\code{generateLpFileCarnival}
}
\details{
Prepare the input data for the run: tranforms data into lp file and .Rdata file.
These files can be reused to run CARNIVAL without preprocessing step using runCarnivalFromLp(..)
}
