% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesymbol2uniprot.R
\name{genesymbol2uniprot}
\alias{genesymbol2uniprot}
\title{\code{genesymbol2uniprot}}
\usage{
genesymbol2uniprot(df, map, geneID, uniprotID)
}
\arguments{
\item{map}{A dataframe with a uniprot ID column and a genesymbol column.}

\item{geneID}{Column index of the gene symbol column in map.}

\item{uniprotID}{Column index of the uniprot column in map.}

\item{names}{A vector of genes.}
}
\value{
A vector of genes or uniprot IDs.
}
\description{
This function converts the gene symbol rownames of a dataframe to uniprot rownames. In case one gene symbol maps to two uniprot IDs, the row will be duplicated.
}
