% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{getOptionsList}
\alias{getOptionsList}
\title{Returns the list of options needed/supported for each solver.}
\usage{
getOptionsList(solver = "", onlyRequired = F)
}
\arguments{
\item{solver}{one of the solvers available from getSupportedSolvers()}

\item{onlyRequired}{logic, set to TRUE if you want to obtain only
required options for the run}
}
\value{
list of options, solver-dependent
}
\description{
Returns the list of options needed/supported for each solver.
}
