% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{getSupportedSolvers}
\alias{getSupportedSolvers}
\title{Returns the list of supported solvers.}
\usage{
getSupportedSolvers()
}
\value{
list of currently supported solvers.
}
\description{
Returns the list of supported solvers.
}
