% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{getSupportedSolversFunctions}
\alias{getSupportedSolversFunctions}
\title{Supported solvers functions to work with all solvers in a uniform way.}
\usage{
getSupportedSolversFunctions()
}
\value{
list of solvers and their corresponding functions.
}
\description{
To add a new solver, one must write and add here the functions for 
3 steps: solve, obtaining a solution matrix, exporting the solution
matrix. More specific functions can be written and called (e.g. check 
saveDiagnostics in cplex).
}
\keyword{internal}
