% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_CARNIVAL_examples.R
\name{load_CARNIVAL_examples}
\alias{load_CARNIVAL_examples}
\title{\code{load_CARNIVAL_examples}}
\usage{
load_CARNIVAL_examples(CARNIVAL_example)
}
\arguments{
\item{CARNIVAL_example}{The number of CARNIVAL example to be loaded; 1 = Toy model, 2 = EGF-SBVimprover,3 = APAP-TGGs}
}
\value{
The input variables for CARNIVAL including network, measurement, input target +/- pathways scores
}
\description{
\code{load_CARNIVAL_examples}
}
