% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapUniprotPKN.R
\name{mapUniprotPKN}
\alias{mapUniprotPKN}
\title{\code{mapUniprotPKN}}
\usage{
mapUniprotPKN(netFile = NULL, organism = "human")
}
\arguments{
\item{netFile}{Prior knowledge network file in Uniprot format}

\item{organism}{organism that the Uniprot IDs refer to ('human','mouse','rat'; default='human')}
}
\value{
An object containing all PKN nodes (universe) in Gene Symbol format written to a file for enrichment analyses
}
\description{
This function generate a file containing all nodes in the prior knowledge network with UniprotID to be assigned as universe for enrichment analysis
}
