% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{prepareForCarnivalRun}
\alias{prepareForCarnivalRun}
\title{Prepares ILP formulation and writes it to .lp file. 
Currently supports the old data representation (CARNIVAL v.<2) 
for debugging and testing if any problems arise with the new way 
to generate variables.}
\usage{
prepareForCarnivalRun(
  dataPreprocessed,
  carnivalOptions,
  newDataRepresentation = T
)
}
\arguments{
\item{dataPreprocessed}{list containing preprocessed priorKnowledgeNetwork, 
measurements, weights (if provided), perturbations (if provided).}

\item{carnivalOptions}{all options of CARNIVAL.}

\item{newDataRepresentation}{TRUE by default. For debugging with the old 
data representation, put to FALSE.}
}
\value{
list with all variables and ILP formulation written in .lp file.
}
\description{
Prepares ILP formulation and writes it to .lp file. 
Currently supports the old data representation (CARNIVAL v.<2) 
for debugging and testing if any problems arise with the new way 
to generate variables.
}
\keyword{internal}
