% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prior_knowledge_network.R
\name{preprocessPriorKnowledgeNetwork}
\alias{preprocessPriorKnowledgeNetwork}
\title{Preprocesses prior knowledge network: correct nodes identifiers for symbols 
that might break solvers runs, assigns the types for each column: 
Node1 (character), Sign (numeric), Node2 (character). 
Stops if interaction/sign column has non-numeric value}
\usage{
preprocessPriorKnowledgeNetwork(priorKnowledgeNetwork)
}
\arguments{
\item{priorKnowledgeNetwork}{a network with 3 columns: 
source node ('source'), interaction 
sign ('interaction') and target node('target').}
}
\value{
preprocessed prior knowledge network with corrected nodes identifiers
add 3 columns: Node1, Sign, Node2
}
\description{
Preprocesses prior knowledge network: correct nodes identifiers for symbols 
that might break solvers runs, assigns the types for each column: 
Node1 (character), Sign (numeric), Node2 (character). 
Stops if interaction/sign column has non-numeric value
}
\author{
Enio Gjerga, Olga Ivanova 2020-2021
}
\keyword{internal}
