% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{processSolution}
\alias{processSolution}
\title{Exports the solution matrix to the final solution.}
\usage{
processSolution(
  solutionMatrix,
  variables,
  dataPreprocessed,
  carnivalOptions,
  newDataRepresentation = T
)
}
\arguments{
\item{solutionMatrix}{the output matrix from ILP solver containing 
variables list (rows) and their values in different solutions (columns).}

\item{variables}{list of nodes, edges and measurements variables 
generated by createLpFormulation_v2.}

\item{dataPreprocessed}{list containing preprocessed priorKnowledgeNetwork, 
measurements, weights (if provided), perturbations (if provided).}

\item{carnivalOptions}{all options of CARNIVAL.}

\item{newDataRepresentation}{TRUE by default. For debugging with the old 
data representation, put to FALSE.}
}
\description{
Exports the solution matrix to the final solution.
}
\keyword{internal}
