% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{readOptions}
\alias{readOptions}
\title{Reads options from json file.}
\usage{
readOptions(jsonFileName = "inst/carnival_cplex_parameters.json")
}
\arguments{
\item{jsonFileName}{path to json files with setups for the solver}
}
\value{
full list of options
}
\description{
Reads options from json file.
}
\keyword{"inst/carnival_cplex_parameters.json")}
\keyword{#TODO}
\keyword{#examples}
\keyword{#readOptions(jsonFileName}
\keyword{=}
\keyword{internal}
