% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dorothea.R
\name{run_dorothea}
\alias{run_dorothea}
\title{\code{run_dorothea}}
\usage{
run_dorothea(df, regulon, confidence_level = c("A", "B", "C"),
  write2file = NULL)
}
\arguments{
\item{df}{A n*m dataframe describing the contrast t-statistics, where n is the number of genes and m the number of conditions.}

\item{regulon}{DoRothEA regulon list}

\item{confidence_level}{A vector describing which confidence levels of the DoRothEA regulon to use}

\item{write2file}{Path to outputfile for all TF_activities if desired}
}
\value{
A n*m dataframe describing the inferred TF activities, where n is the number of Tfs and m the number of conditions.
}
\description{
This function infers TF activities based on the DoRothEA regulon and the viper function.
The regulon can be filtered by confidence level (A-E).
}
