% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_run_carnival.R
\name{runCarnivalFromLp}
\alias{runCarnivalFromLp}
\title{\code{runCarnivalFromLp}}
\usage{
runCarnivalFromLp(
  lpFile = "",
  parsedDataFile = "",
  solver = supportedSolvers$lpSolve,
  solverPath = "",
  carnivalOptions = defaultLpSolveCarnivalOptions()
)
}
\arguments{
\item{perturbations}{vector of targets of perturbations.}

\item{measurements}{vector of the measurements (i.e. DoRothEA/VIPER normalised
enrichment scores)}

\item{priorKnowledgeNetwork}{data frame of the prior knowledge network}

\item{pathwayWeights}{(optional) vector of the additional weights: e.g. PROGENy pathway
score or measured protein activities.

#'@return data frame of all variables for ILP formulation.
#'@return The function will return a list of results containing:
1. weightedSIF: A table with 4 columns containing the combined network
solutions from CARNIVAL. It contains the Source of the interaction (Node1),
Sign of the interaction (Sign), the Target of the interaction (Node2) and the
weight of the interaction (Weight) which shows how often an interaction
appears across all solutions.

2. nodesAttributes: A table with 6 columns containing information about
infered protein activity states and attributes. It contains the Protein IDs
(Node); how often this node has taken an activity of 0, 1 and -1 across the
solutions (ZeroAct, UpAct, DownAct); the average activities across solutions
(AvgAct); and the node attribute (measured, target, inferred).

3. sifAll: A list of separate network solutions.

4. attributesAll: A list of separate inferred node activities in each
solution.

5. diagnostics: reports the convergence of optimization and reason of 
the termination. Only for CPLEX solver.}
}
\description{
\code{runCarnivalFromLp}
}
\details{
Runs CARNIVAL pipeline with preparsed data - lp file and Rdata file containing variables for ILP formulation.
}
\author{
Enio Gjerga, Olga Ivanova 2020-2021 \email{carnival.developers@gmail.com}
}
