% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_run_carnival.R
\name{runInverseCarnival}
\alias{runInverseCarnival}
\title{\code{runInverseCarnival}}
\usage{
runInverseCarnival(
  measurements,
  priorKnowledgeNetwork,
  pathwayWeights = NULL,
  solverPath = solverPath,
  solver = supportedSolvers$lpSolve,
  carnivalOptions = defaultCplexCarnivalOptions(solverPath = solverPath)()
)
}
\arguments{
\item{lpFile}{full path to .lp file}

\item{parsedDataFile}{full path to preprocessed .RData file

1. weightedSIF: A table with 4 columns containing the combined network
solutions from CARNIVAL. It contains the Source of the interaction (Node1),
Sign of the interaction (Sign), the Target of the interaction (Node2) and the
weight of the interaction (Weight) which shows how often an interaction
appears across all solutions.

2. nodesAttributes: A table with 6 columns containing information about
infered protein activity states and attributes. It contains the Protein IDs
(Node); how often this node has taken an activity of 0, 1 and -1 across the
solutions (ZeroAct, UpAct, DownAct); the average activities across solutions
(AvgAct); and the node attribute (measured, target, inferred).

3. sifAll: A list of separate network solutions.

4. attributesAll: A list of separate inferred node activities in each
solution.

5. diagnostics: reports the convergence of optimization and reason of
the termination. Only for CPLEX solver.}
}
\description{
\code{runInverseCarnival}
}
\details{
Runs CARNIVAL pipeline with preparsed data - lp file and Rdata file containing variables for ILP formulation.
}
\author{
Enio Gjerga, Olga Ivanova 2020-2021 \email{carnival.developers@gmail.com}
}
