% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPROGENy.R
\name{runPROGENy}
\alias{runPROGENy}
\title{\code{runPROGENy}
This function is designed to compute progeny pathway scores and assess there significance using a gene sampling based permutation strategy, for a series of experimental samples/contrasts.}
\usage{
runPROGENy(df, weight_matrix, k = 10000, z_scores = T,
  get_nulldist = F)
}
\arguments{
\item{df}{A data.frame of n*m+1 dimension, where n is the number of omic features to be considered and m is the number of samples/contrasts.
The first column should be the identifiers of the omic features. These identifiers must be coherent with the identifers of the weight matrix.}

\item{weight_matrix}{A progeny coeficient matrix. the first column should be the identifiers of the omic features, and should be coherent with the identifiers provided in df.}

\item{k}{The number of permutations to be preformed to generate the null-distribution used to estimate significance of progeny scores. Default value is 10000.}

\item{z_scores}{If true, provides z-scores. If false, provides significance scores (1-pval).}

\item{get_nulldist}{If get_nulldist is true, then the function will return the null model dataframe that was used.}
}
\value{
This function returns a list of two elements. The first element is a dataframe of p*m+1 dimensions, where p is the number of progeny pathways, and m is the number of samples/contrasts.
Each cell represent the significance of a progeny pathway score for one sample/contrast. The signifcance ranges between -1 and 1. The significance is equal to x*2-1, x being the quantile of the progeny pathway score with respect to the null distribution.
Thus, this significance can be interpreted as the equivalent of 1-p.value (two sided test over an empirical distribution) with the sign indicating the direction of the regulation.
The sceond element is the null distribution list (a null distribution is generated for each sample/contrast).
}
\description{
\code{runPROGENy}
This function is designed to compute progeny pathway scores and assess there significance using a gene sampling based permutation strategy, for a series of experimental samples/contrasts.
}
