% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_solver_path.R
\name{searchForCPLEXSolver}
\alias{searchForCPLEXSolver}
\title{searchForCPLEXSolver}
\usage{
searchForCPLEXSolver()
}
\value{
path to CPLEX if found or empty string ("") if not.
}
\description{
search for interactive cplex executable in the default path
}
