% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{sendTaskToSolver}
\alias{sendTaskToSolver}
\title{Executes the solve on the provided ILP formulation (in .lp file).}
\usage{
sendTaskToSolver(
  variables,
  dataPreprocessed,
  carnivalOptions,
  newDataRepresentation = T
)
}
\arguments{
\item{variables}{list of nodes, edges and measurements variables 
generated by createLpFormulation_v2.}

\item{dataPreprocessed}{list containing preprocessed priorKnowledgeNetwork, 
measurements, weights (if provided), perturbations (if provided).}

\item{carnivalOptions}{all options of CARNIVAL.}

\item{newDataRepresentation}{TRUE by default. For debugging with the old 
data representation, put to FALSE.}
}
\value{
solution matrix from ILP solver containing variables list (rows) and 
their values in different solutions (columns).
}
\description{
Executes the solve on the provided ILP formulation (in .lp file).
}
\keyword{internal}
