% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{setCarnivalOptions}
\alias{setCarnivalOptions}
\title{Sets CARNIVAL options for the solver.}
\usage{
setCarnivalOptions(solver = getSupportedSolvers()$lpSolve, ...)
}
\arguments{
\item{solver}{one of the solvers available from getSupportedSolvers().}

\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Sets CARNIVAL options for the solver.
}
\examples{
setCarnivalOptions(solver="lpSolve")
}
