% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCARNIVAL.R
\name{shinyCARNIVAL}
\alias{shinyCARNIVAL}
\title{\code{shinyCARNIVAL}}
\usage{
shinyCARNIVAL(dir_name, nrIntActInc = 1000, thresNodeAll = NULL,
  thresEdgeAll = NULL, graphLayout = "concentric", UP2GS = F,
  inverseCR = F)
}
\arguments{
\item{dir_name}{The directory name that stores CARNIVAL results}

\item{nrIntActInc}{Maximum number of included interactions in the graph}

\item{thresNodeAll}{Minimal node activity to be shown on the graph}

\item{thresEdgeAll}{Minimal edge presence percentage to be shown on the graph}

\item{graphLayout}{Defined graph layout: "cose", "breadthfirst", "concentric", "preset"}

\item{UP2GS}{Convert UniprotID to Gene Symbols to facilitate the interpretation}

\item{inverseCR}{Set parameter to TRUE to remove the perturbation node from the network upon display}
}
\value{
Launched Rshiny app with plotted results
}
\description{
This function takes results from ILP optimisation and display the results on an R-Shiny application
}
