% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{solveCarnival}
\alias{solveCarnival}
\title{Main CARNIVAL function to execute the full pipeline: 
1) preprocess the data
2) prepare ILP formulation
3) executes the solver on ILP formulation 
4) parse the output of the solver and map it to the original data.}
\usage{
solveCarnival(dataPreprocessed, carnivalOptions, newDataRepresentation = T)
}
\arguments{
\item{dataPreprocessed}{list containing preprocessed priorKnowledgeNetwork, 
measurements, weights (if provided), perturbations (if provided).}

\item{carnivalOptions}{all options of CARNIVAL.}

\item{newDataRepresentation}{TRUE by default. For debugging with the old 
data representation, put to FALSE.}
}
\value{
solution of the ILP problem.
}
\description{
Main CARNIVAL function to execute the full pipeline: 
1) preprocess the data
2) prepare ILP formulation
3) executes the solver on ILP formulation 
4) parse the output of the solver and map it to the original data.
}
\keyword{internal}
