% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{solveCarnivalFromLp}
\alias{solveCarnivalFromLp}
\title{Sends the ILP formulation defined in .lp file to solver. Uses parsedDataFile
to process the final solution and map the ILP variables back to inital 
data.}
\usage{
solveCarnivalFromLp(
  lpFile = "",
  parsedDataFile = "",
  carnivalOptions,
  newDataRepresentation = T
)
}
\arguments{
\item{lpFile}{path to .lp file that will be used to run the solver.}

\item{parsedDataFile}{path to parsed data file that was created after running 
\code{\link{prepareForCarnivalRun}} or in previous CARNIVAL runs.}

\item{carnivalOptions}{all options of CARNIVAL.}

\item{newDataRepresentation}{TRUE by default. For debugging with the old 
data representation, put to FALSE.}
}
\value{
solution of ILP problem
}
\description{
Sends the ILP formulation defined in .lp file to solver. Uses parsedDataFile
to process the final solution and map the ILP variables back to inital 
data.
}
\keyword{internal}
