% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers_cbc__solve_cbc.R
\name{solveWithCbc}
\alias{solveWithCbc}
\title{Executes cbc solver on provided .lp file.}
\usage{
solveWithCbc(carnivalOptions)
}
\arguments{
\item{carnivalOptions}{}
}
\value{

}
\description{
Executes cbc solver on provided .lp file.
}
\keyword{internal}
