% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers_gurobi__solve_gurobi.R
\name{solveWithGurobi}
\alias{solveWithGurobi}
\title{Executes gurobi solver on provided .lp file.}
\usage{
solveWithGurobi(carnivalOptions)
}
\arguments{
\item{carnivalOptions}{}
}
\value{
Returns the name of the result files without ".sol" extension.
}
\description{
Executes gurobi solver on provided .lp file.
}
\keyword{internal}
