% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{suggestedCplexSpecificOptions}
\alias{suggestedCplexSpecificOptions}
\title{Suggests cplex specific options.s}
\usage{
suggestedCplexSpecificOptions(...)
}
\arguments{
\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Suggests cplex specific options.s
}
\examples{
suggestedCplexSpecificOptions()
}
