% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_solver_path.R
\name{testRunCplex}
\alias{testRunCplex}
\title{checks if the the cplexPath exists}
\usage{
testRunCplex(cplexPath)
}
\arguments{
\item{cplexPath}{path to cplex to check}
}
\value{
TRUE if interaction was successful, FALSE otherwise
}
\description{
tries to interact with the cplex
}
