% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uniprot2GeneSymbol.R
\name{Uniprot2GeneSymbol}
\alias{Uniprot2GeneSymbol}
\title{\code{Uniprot2GeneSymbol}}
\usage{
Uniprot2GeneSymbol(res)
}
\arguments{
\item{res}{A list of variables in Uniprot ID (network and node activities)}
}
\value{
A list of mapped variables from Uniprot to official gene symbol for a better readibility
}
\description{
Conversion of Uniprot ID (e.g. from Omnipath) to official gene symbol in the plotting step
}
