% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_binaries.R
\name{write_binaries}
\alias{write_binaries}
\title{\code{write_binaries}}
\usage{
write_binaries(variables = variables)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}
}
\value{
This code writes the list of binary variables (xp, xm, up & um).
}
\description{
\code{write_binaries}
}
