% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_boundaries.R
\name{write_boundaries}
\alias{write_boundaries}
\title{\code{write_boundaries}}
\usage{
write_boundaries(variables = variables, oF = oF)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}

\item{oF}{Is the objective function of the formulation.}
}
\value{
This code writes the boundaries of each variable.
}
\description{
\code{write_boundaries}
}
