% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_constraints_5_all.R
\name{write_constraints_5_all}
\alias{write_constraints_5_all}
\title{\code{write_constraints_5_all}}
\usage{
write_constraints_5_all(variables = variables)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}
}
\value{
This code writes the list of constraints (5) of the ILP problem for all the conditions.
}
\description{
\code{write_constraints_5_all}
}
