% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_constraints_7.R
\name{write_constraints_7}
\alias{write_constraints_7}
\title{\code{write_constraints_7}}
\usage{
write_constraints_7(variables = variables, dataMatrix = dataMatrix,
  inputs = inputs)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}

\item{dataMatrix}{Contains the matrix which stores the data for running CARNIVAL and a set of identifiers for Targets, Measured and Un-measured nodes.}

\item{inputs}{Contains the list of targets as inputs.}
}
\value{
This code writes the list of constraints (7) of the ILP problem for all the conditions.
}
\description{
\code{write_constraints_7}
}
