% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_constraints_8.R
\name{write_constraints_8}
\alias{write_constraints_8}
\title{\code{write_constraints_8}}
\usage{
write_constraints_8(variables = variables, inputs = inputs,
  pknList = pknList)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}

\item{inputs}{Contains the list of targets as inputs.}

\item{pknList}{Contains the background network which serves as a prior knowledge and which we train.}
}
\value{
This code writes the list of constraints (8) of the ILP problem for all the conditions.
}
\description{
\code{write_constraints_8}
}
