% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_constraints_objFunction_all.R
\name{write_constraints_objFunction_all}
\alias{write_constraints_objFunction_all}
\title{\code{write_constraints_objFunction_all}}
\usage{
write_constraints_objFunction_all(variables = variables,
  dataMatrix = dataMatrix)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indeces.}

\item{dataMatrix}{Contains the matrix which stores the data for running CARNIVAL and a set of identifiers for Targets, Measured and Un-measured nodes.}
}
\value{
This function returns the list of constraints associated with the 'Absolute Difference' variables and which measure the mis-fit between inferred and measured data.
}
\description{
\code{write_constraints_objFunction_all}
}
