% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_generals.R
\name{write_generals}
\alias{write_generals}
\title{\code{write_generals}}
\usage{
write_generals(variables = variables, oF = oF)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}

\item{oF}{Is the objective function of the formulation.}
}
\value{
This code writes all the variables.
}
\description{
\code{write_generals}
}
