% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_loop_constraints.R
\name{write_loop_constraints}
\alias{write_loop_constraints}
\title{\code{write_loop_constraints}}
\usage{
write_loop_constraints(variables = variables, pknList = pknList,
  inputs = inputs)
}
\arguments{
\item{variables}{Contains the list of variables as used to formulate the ILP problem, explanations for each variable and a list of useful indices.}

\item{pknList}{Contains the background network which serves as a prior knowledge and which we train.}

\item{inputs}{Contains the list of targets as inputs.}
}
\value{
This function writes the constraints preventing self-activation of nodes in the network due to positive feedback loops.
}
\description{
\code{write_loop_constraints}
}
