% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers_cplex__write_cplex_command_file.R
\name{writeCplexCommandFileFromJson}
\alias{writeCplexCommandFileFromJson}
\title{writeCplexCommandFileFromJson}
\usage{
writeCplexCommandFileFromJson(
  carnivalOptions,
  jsonFileName = "parameters/cplex_parameters_cmd_file.json"
)
}
\value{
list of params
}
\description{
writeCplexCommandFileFromJson
}
