% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDOTfig.R
\name{WriteDOTfig}
\alias{WriteDOTfig}
\title{\code{WriteDOTfig}}
\usage{
WriteDOTfig(res, idxModel = 0, dir_name, inputs, measurements,
  UP2GS = F)
}
\arguments{
\item{res}{A list containing optimised network structures and node activities from ILP optimisation}

\item{idxModel}{(optional) The index or indices of optimised model to be plotted (Default idxModel=0; plot the combined and average network)}

\item{dir_name}{The directory name that stores CARNIVAL results (as a subfolder of the main "results" folder)}

\item{inputs}{A named vector of inputs and their node states}

\item{measurements}{A named vector of measurements and their state values}

\item{UP2GS}{(T/F) A parameter defining whether to convert the input format UniprotID to Gene Symbol}
}
\value{
DOT figures stored in the directory "results/dir_name" with the prefix "ActivityNetwork_model"
}
\description{
This function takes results from ILP optimisation and write out a figure in DOT format
}
