% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLPFile.R
\name{writeLPFile}
\alias{writeLPFile}
\title{\code{writeLPFile}}
\usage{
writeLPFile(data = data, pknList = pknList, inputs = inputs,
  alphaWeight = 1, betaWeight = 0.2, scores = scores, mipGAP = 0.1,
  poolrelGAP = 0.01, limitPop = 100, poolCap = 100,
  poolIntensity = 0, poolReplace = 2, timelimit = 1800,
  measWeights = NULL, repIndex, condition = "")
}
\arguments{
\item{data}{The discretised observations (here transcription factor activities) of values [-1,0,1]}

\item{pknList}{The prior knowledge network}

\item{inputs}{The list of known or potential target of perturbation}

\item{alphaWeight}{The weight for mismatch penalty between discretised observations and predicted model states}

\item{betaWeight}{The weight for network size (node) penalty}

\item{scores}{The continuous pathway scores from PROGENy}

\item{mipGAP}{The minimal integer programming percentage gap to be accepted as a solution}

\item{poolrelGAP}{The allowed relative percentage gap between the best solution and the equivalent solutions in the solution pool}

\item{limitPop}{The number of allowed solutions to be populated}

\item{poolCap}{The number of solutions to be kept in the pool of solution}

\item{poolIntensity}{The intensity of the search in solution space}

\item{poolReplace}{The replacement strategy of the solutions in the solution pool}

\item{timelimit}{The allowed amount of time (in seconds) for the optimisation}

\item{measWeights}{The countinous weight of observations (here transcription factor activities) - to replace the default alphaWeight if assigned; Note: take only positive values!}

\item{repIndex}{The indexing of optimisation - useful in case more than one experiment is performed}

\item{condition}{The free variable which could be assigned for additional study e.g. to vary the efect of betaWeight in a loop}
}
\value{
An integer programming file containing the description of ILP optimisation problem and a cplex command file to communicate with the interactive version of cplex solver
}
\description{
Write a list of linear programming (LP) constraints into a file while will be read by interactive cplex solver to perform network optimisation.
}
