% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_carnival.R
\name{writeParsedData}
\alias{writeParsedData}
\title{Saves all provided data together with generated variables for ILP problem in 
.RData file.}
\usage{
writeParsedData(
  variables = variables,
  dataPreprocessed = dataPreprocessed,
  filename = "parsedData.RData"
)
}
\arguments{
\item{variables}{list of nodes, edges and measurements variables 
generated by createLpFormulation_v2}

\item{dataPreprocessed}{list containing preprocessed priorKnowledgeNetwork, 
measurements, weights (if provided), perturbations (if provided).}

\item{filename}{filename of the parsed data file.}
}
\value{
filename of the parsed data file.
}
\description{
Saves all provided data together with generated variables for ILP problem in 
.RData file.
}
\keyword{internal}
