% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{defaultCplexCarnivalOptions}
\alias{defaultCplexCarnivalOptions}
\title{Sets default CARNIVAL options for cplex.}
\usage{
defaultCplexCarnivalOptions(...)
}
\arguments{
\item{...}{any possible options from the solver's list}
}
\description{
Sets default CARNIVAL options for cplex.
}
