% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichCARNIVAL.R
\name{enrichCARNIVAL}
\alias{enrichCARNIVAL}
\title{\code{enrichCARNIVAL}}
\usage{
enrichCARNIVAL(Result_dir = "Results_CARNIVAL", universeFile = NULL,
  networkID = 1, universeID = 2, datapath = NULL,
  datasource = NULL, directionalORA = T, undirectionalORA = T,
  plot = F, topEnrichAll = 40, pathwayfilter = T, pValSig = 0.05)
}
\arguments{
\item{Result_dir}{The Result CARNIVAL folder to perform enrichment analyses on}

\item{universeFile}{The file containing the list of nodes/genes (with the header(s) 'genesymbol' +/- 'uniprot') to be considered as universe in the enrichment analyses}

\item{networkID}{Index of the column with the network identifiers used in the universeFile}

\item{universeID}{Index of the column with the genesymbol identifiers in the universeFile}

\item{datapath}{Path to GMT file if not specified by datasource}

\item{datasource}{The data source on MSigDB C2 branch: c("kegg":default,"biocarta","reactome","allc2","cpg","cp","hallmark")}

\item{directionalORA}{If TRUE directional over-representation analysis is performed}

\item{undirectionalORA}{If TRUE undirectional over-representation analysis is performed}

\item{plot}{If TRUE plots are generated}

\item{topEnrichAll}{Number of top enrichmed pathways to be plotted (default: 40)}

\item{pathwayfilter}{Filtering out pathways which are related to cancers, infections and diseases to highlight cellular processes}

\item{pValSig}{Significant p-value to be labelled on the plotted figures}
}
\value{
Enriched pathways with p-value written into files and figures showing top enriched pathways for combined Up/Down direction and separated
}
\description{
This function performs the enrichment of CARNIVAL nodes with over-representation analysis using the curated gene set from MSigDB (C2 branch)
The enrichment can be performed using the KEGG, Biocarta or Reactome subsets or using the whole curated C2 geneset.
Note: The universe of the geneset should be all the nodes in the prior knowledge network. Here the nodes from Omnipath are used by default.
}
