% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSymbol2Uniprot.R
\name{GeneSymbol2Uniprot}
\alias{GeneSymbol2Uniprot}
\title{\code{GeneSymbol2Uniprot}}
\usage{
GeneSymbol2Uniprot(df, map, geneID = 1, uniprotID = 2)
}
\arguments{
\item{df}{A vector of genes.}

\item{map}{A dataframe with a uniprot ID column and a genesymbol column.}

\item{geneID}{Column index of the gene symbol column in map.}

\item{uniprotID}{Column index of the uniprot column in map.}
}
\value{
A vector of genes or uniprot IDs.
}
\description{
This function converts the gene symbol rownames of a dataframe to uniprot rownames. In case one gene symbol maps to two uniprot IDs, the row will be duplicated.
}
