% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_CPLEX_log.R
\name{parse_CPLEX_log}
\alias{parse_CPLEX_log}
\title{parse carnival logs}
\usage{
parse_CPLEX_log(log_file)
}
\arguments{
\item{log_file}{path of log file resulted from a carnival run}
}
\value{
list variable with following fields:
- `convergence` a table that contains information on the convergence of CPLEX 
- `n_solutions` number of solutions found 
- `objective` objective function value
- `termination_reason`: reason of termination.
}
\description{
parses the CPLEX log file and reads some basic information
}
