% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCARNIVAL.R
\name{runCARNIVAL}
\alias{runCARNIVAL}
\title{\code{runCARNIVAL}}
\usage{
runCARNIVAL(InputFile = "InputFile_CARNIVAL.txt")
}
\arguments{
\item{InputFile}{The filename of input file for CARNIVAL with a fixed structure (default see: InputFile_CARNIVAL.txt)}
}
\value{
The networks and predicted node activities from the CARNIVAL pipeline in the destined result folder
}
\description{
Run CARNIVAL pipeline using to the provided list of inputs (including CARNIVAL built-in examples)
}
