% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_carnival_options.R
\name{suggestedCbcSpecificOptions}
\alias{suggestedCbcSpecificOptions}
\title{Suggests cbc specific options.}
\usage{
suggestedCbcSpecificOptions(...)
}
\arguments{
\item{...}{any possible options from the solver's list}
}
\value{

}
\description{
Suggests cbc specific options.
}
\examples{
suggestedCbcSpecificOptions()
}
